Django/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
investpy==1.0.8
jiter==0.7.1
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.54.4
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
pyupbit==0.2.34
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
statsmodels==0.14.4
textblob==0.18.0.post0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
websockets==14.1
zipp==3.21.0
